Default File Changes Often
It seems that opening the Unity editor leads to the default-2021 file changing. Perhaps, the is brought about by the movement of the scene or "reload script assemblies". Whatever the case, we do not need to worry about it. It is a procedure that Unity runs and does not pertain to the developer