Bugs fixes and Changes toNetworkedEnemyFollow
Changes to Trampoline
Changed to NetworkedEnemyFollow
Some prefab changes.
Fixed the new longtime coinsScripts warning that had "new if hiding was intended."
Fixed some bugs showing up in the console
Added ConnectionStatus script that and a textmeshpro on the master computer prefab. It tells the player if they are connected to a photon server or not.